requirements.txt
tensorflow-gpu
tqdm
pypiwin32
nltk
pygame
numpy
SpeechRecognition
gtts
pyttsx3
word2number
playsound
mutagen
transformers
pylogix
gTTS
pydub
soco